requirements.txt
backtrader
alpaca-trade-api
pandas
matplotlib
python-dotenv
